requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.0.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
hiredis==2.2.3
Pillow==10.0.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
